create client.py & mv consumet etc...
