+++ n0core/scheduler/agent/volumer.py
@@ -1,1 +1,7 @@
-from main import consumer, logger, scheduler  # NOQA
+from client import consumer, logger, send  # NOQA
+try:
+    from n0core.lib import proto
+except:  # NOQA
+    import sys
+    sys.path.append('../../')
+    from n0core.lib import proto
